requirements.txt
asgiref==3.4.1
Django==3.0
django-crispy-forms==1.13.0
djangorestframework==3.12.4
django-hitcount==1.3.3
django-taggit==2.1.0
markdown==3.3.4
pytz==2021.3
sqlparse==0.4.2
Pillow==7.0.0